* src/org/jcae/mesh/amibe/ds/HalfEdge.java: Change 'attributes' type from int to byte.
git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@1816 9c4724a3-da1e-0410-bc68-88f18da35cd4

+++ amibe/src/org/jcae/mesh/amibe/ds/HalfEdge.java
@@ -36,7 +36,7 @@ public class HalfEdge extends AbstractHalfEdge implements Serializable
 	private static Logger logger = Logger.getLogger(HalfEdge.class);
 	private TriangleHE tri;
 	private byte localNumber = 0;
-	private int attributes = 0;
+	private byte attributes = 0;
 	private HalfEdge sym;
 	private HalfEdge next;
